Ensure that re-including the snippet is harmless
Often times snippets are accidentally included more than once (maybe
through the app code and something like Google Tag Manager) and we don't
want to re-stub the actual stringer after it has been loaded!

In the future we may want to log that the snippet has been included more
than once so that it can be rectified, but reloading snippets is often
part of a normal development workflow with something like figwheel.