Always use the passed in userEmail and userId in identify
Otherwise, if we've identified the userEmail and userId and then later only
userEmail, we end up in a situation that the userId now has a new email. Even
worse, the cookie didn't have this guarding logic of nulls and so the in memory
and the cookie versions of the values differ (making things even more
confusing). Instead, we just don't need these guards.