Remove sendBeacon and always use XHR requests
It seems that on iOS, webviews/webkit browsers have a broken sendBeacon
implementation (specifically only in-app browsers, like clicking links in the
Facebook app, Instagram app, or Chrome on iOS). Normal Safari does not reproduce
the issue.

It seems like it is because the sendBeacon implementation on Safari is broken,
as per
https://stackoverflow.com/questions/51844586/navigator-sendbeacon-on-iphone-chrome-not-posting-to-secure-websites?rq=1

We'll try this commit to see if it removes the issue, or if there is some other
specific issue that is causing this to fail.